+++ cli/tests/test_git_init.rs
@@ -16,9 +16,13 @@ use std::fmt::Write as _;
 use std::path::Path;
 use std::path::PathBuf;
 
+use indoc::formatdoc;
 use test_case::test_case;
 
+use crate::common::get_stderr_string;
+use crate::common::get_stdout_string;
 use crate::common::strip_last_line;
+use crate::common::to_toml_value;
 use crate::common::TestEnvironment;
 
 fn init_git_repo(git_repo_path: &Path, bare: bool) -> git2::Repository {
@@ -764,6 +768,71 @@ fn test_git_init_colocated_via_flag_git_dir_not_exists() {
     @  230dd059e1b0 main master
     ◆  000000000000
     "###);
+}
+
+#[test]
+fn test_git_init_conditional_config() {
+    let test_env = TestEnvironment::default();
+    let old_workspace_root = test_env.env_root().join("old");
+    let new_workspace_root = test_env.env_root().join("new");
+
+    let jj_cmd_ok = |current_dir: &Path, args: &[&str]| {
+        let mut cmd = test_env.jj_cmd(current_dir, args);
+        cmd.env_remove("JJ_EMAIL");
+        cmd.env_remove("JJ_OP_HOSTNAME");
+        cmd.env_remove("JJ_OP_USERNAME");
+        let assert = cmd.assert().success();
+        let stdout = test_env.normalize_output(&get_stdout_string(&assert));
+        let stderr = test_env.normalize_output(&get_stderr_string(&assert));
+        (stdout, stderr)
+    };
+    let log_template = r#"separate(' ', author.email(), description.first_line()) ++ "\n""#;
+    let op_log_template = r#"separate(' ', user, description.first_line()) ++ "\n""#;
+
+    // Override user.email and operation.username conditionally
+    test_env.add_config(formatdoc! {"
+        user.email = 'base@example.org'
+        operation.hostname = 'base'
+        operation.username = 'base'
+        [[--scope]]
+        --when.repositories = [{new_workspace_root}]
+        user.email = 'new-repo@example.org'
+        operation.username = 'new-repo'
+        ",
+        new_workspace_root = to_toml_value(new_workspace_root.to_str().unwrap()),
+    });
+
+    // Override operation.hostname by repo config, which should be loaded into
+    // the command settings, but shouldn't be copied to the new repo.
+    jj_cmd_ok(test_env.env_root(), &["git", "init", "old"]);
+    jj_cmd_ok(
+        &old_workspace_root,
+        &["config", "set", "--repo", "operation.hostname", "old-repo"],
+    );
+    jj_cmd_ok(&old_workspace_root, &["new"]);
+    let (stdout, _stderr) = jj_cmd_ok(&old_workspace_root, &["op", "log", "-T", op_log_template]);
+    insta::assert_snapshot!(stdout, @r"
+    @  base@old-repo new empty commit
+    ○  base@base add workspace 'default'
+    ○  @
+    ");
+
+    // Create new repo at the old workspace directory.
+    let (_stdout, stderr) = jj_cmd_ok(&old_workspace_root, &["git", "init", "../new"]);
+    insta::assert_snapshot!(stderr.replace('\\', "/"), @r#"Initialized repo in "../new""#);
+    jj_cmd_ok(&new_workspace_root, &["new"]);
+    let (stdout, _stderr) = jj_cmd_ok(&new_workspace_root, &["log", "-T", log_template]);
+    insta::assert_snapshot!(stdout, @r"
+    @  new-repo@example.org
+    ○  new-repo@example.org
+    ◆
+    ");
+    let (stdout, _stderr) = jj_cmd_ok(&new_workspace_root, &["op", "log", "-T", op_log_template]);
+    insta::assert_snapshot!(stdout, @r"
+    @  new-repo@base new empty commit
+    ○  new-repo@base add workspace 'default'
+    ○  @
+    ");
 }
 
 #[test]